Prototype/server/requirements-dev.txt
pylint>=3.0.0
pylint_flask
pytest>=7.4.0
